# internal/apidocs/openapi.yaml
openapi: 3.1.0
info:
  title: Simple Task Manager API
  version: 1.5.0
  description: |
    REST API for user authentication and per-user task management.
    Protected task endpoints require a Bearer JWT access token.
    This document is the governed public compatibility contract for API v1.
...
servers:
  - url: http://localhost:8080
    description: Local development
...
paths:
...
  /api/auth/register:
    post:
      tags: [Authentication]
      summary: Register a user
      operationId: register
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: User registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthEnvelope'
        '400':
          $ref: '#/components/responses/BadRequest'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '409':
          description: Email already registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '429':
          $ref: '#/components/responses/TooManyRequests'
  /api/auth/login:
    post:
      tags: [Authentication]
      summary: Login
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthEnvelope'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '429':
          $ref: '#/components/responses/TooManyRequests'
  /api/auth/refresh:
    post:
      tags: [Authentication]
      summary: Rotate a refresh token
      operationId: refreshToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenRequest'
      responses:
        '200':
          description: Access and refresh tokens rotated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthEnvelope'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/auth/logout:
    post:
      tags: [Authentication]
      summary: Revoke a refresh token
      operationId: logout
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenRequest'
      responses:
        '200':
          description: Refresh token revoked
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageEnvelope'
        '400':
          $ref: '#/components/responses/BadRequest'
  /api/auth/forgot-password:
    post:
      tags: [Authentication]
      summary: Request password reset instructions
      operationId: forgotPassword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ForgotPasswordRequest'
      responses:
        '200':
          description: Generic password reset response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActionTokenEnvelope'
        '429':
          $ref: '#/components/responses/TooManyRequests'
  /api/auth/reset-password:
    post:
      tags: [Authentication]
      summary: Reset password with a one-time token
      operationId: resetPassword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResetPasswordRequest'
      responses:
        '200':
          description: Password reset
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageEnvelope'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/TooManyRequests'
  /api/auth/change-password:
    post:
      tags: [Authentication]
      summary: Change the current user's password
      operationId: changePassword
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangePasswordRequest'
      responses:
        '200':
          description: Password changed and refresh sessions revoked
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageEnvelope'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/TooManyRequests'
  /api/auth/logout-all:
    post:
      tags: [Authentication]
      summary: Revoke all refresh sessions for the current user
      operationId: logoutAll
      security:
        - bearerAuth: []
      responses:
        '200':
          description: All refresh sessions revoked
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageEnvelope'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/auth/sessions:
    get:
      tags: [Authentication]
      summary: List active refresh sessions
      operationId: listSessions
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Active sessions ordered by last use
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionListEnvelope'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/auth/email-verification/request:
    post:
      tags: [Authentication]
      summary: Request an email verification token
      operationId: requestEmailVerification
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Verification instructions created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActionTokenEnvelope'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/TooManyRequests'
  /api/auth/email-verification/confirm:
    post:
      tags: [Authentication]
      summary: Verify email with a one-time token
      operationId: verifyEmail
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VerifyEmailRequest'
      responses:
        '200':
          description: Email verified
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageEnvelope'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/TooManyRequests'
  /api/workspaces:
    get:
      tags: [Workspaces]
      summary: List workspaces visible to the current user
      operationId: listWorkspaces
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Workspace memberships
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkspaceAccessListEnvelope'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      tags: [Workspaces]
      summary: Create a shared workspace
      operationId: createWorkspace
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateWorkspaceRequest'
      responses:
        '201':
          description: Workspace created with current user as owner
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkspaceAccessEnvelope'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/workspaces/{id}/webhooks:
    parameters:
      - $ref: '#/components/parameters/WorkspaceID'
    get:
      tags: [Webhooks]
      summary: List webhook subscriptions for a workspace
      operationId: listWebhookSubscriptions
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Webhook subscriptions without signing secrets
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebhookSubscriptionListEnvelope'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/WorkspaceNotFound'
    post:
      tags: [Webhooks]
      summary: Create a webhook subscription
      description: The signing_secret is returned only in this creation response.
      operationId: createWebhookSubscription
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateWebhookSubscriptionRequest'
      responses:
        '201':
          description: Webhook subscription created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebhookSubscriptionEnvelope'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/WorkspaceNotFound'
  /api/workspaces/{id}/webhooks/{subscription_id}:
    parameters:
      - $ref: '#/components/parameters/WorkspaceID'
      - $ref: '#/components/parameters/WebhookSubscriptionID'
    delete:
      tags: [Webhooks]
      summary: Delete a webhook subscription
      operationId: deleteWebhookSubscription
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Webhook subscription deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageEnvelope'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          description: Workspace or webhook subscription not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /api/tasks:
    parameters:
      - $ref: '#/components/parameters/WorkspaceIDHeader'
    get:
      tags: [Tasks]
      summary: List current user's tasks
      operationId: listTasks
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            minimum: 1
            default: 1
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
        - in: query
          name: search
          schema:
            type: string
          description: Case-insensitive search in title and description.
        - in: query
          name: completed
          schema:
            type: boolean
        - in: query
          name: sort
          schema:
            type: string
            enum: [id, title, created_at, updated_at, completed]
            default: created_at
        - in: query
          name: order
          schema:
            type: string
            enum: [asc, desc]
            default: desc
      responses:
        '200':
          description: Paginated task list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskPageEnvelope'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      tags: [Tasks]
      summary: Create a task
      operationId: createTask
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskWriteRequest'
      responses:
        '201':
          description: Task created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskEnvelope'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          description: Idempotency key conflict or another request with the same key is still in progress
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
...
  /api/oauth/token:
    post:
      tags: [Enterprise Identity]
      summary: Exchange an OAuth2 authorization code or client credentials for an access token
      operationId: oauthToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [grant_type, client_id, client_secret]
              additionalProperties: true
      responses:
        '200':
          description: OAuth2 access token issued
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/oauth/api-key:
    post:
      tags: [Enterprise Identity]
      summary: Exchange a workspace API key for a scoped service access token
      operationId: exchangeServiceApiKey
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [api_key]
              properties:
                api_key:
                  type: string
      responses:
        '200':
          description: Scoped service access token issued
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    Unauthorized:
      description: Missing, invalid, or expired access token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
...
    TooManyRequests:
      description: Authentication rate limit exceeded
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
  schemas:
    RegisterRequest:
      type: object
      required: [name, email, password]
      properties:
        name:
          type: string
          minLength: 1
          example: Yudi
        email:
          type: string
          format: email
          example: yudi@example.com
        password:
          type: string
          minLength: 8
          format: password
          example: password123
...
    ErrorEnvelope:
      type: object
      required: [success, message]
      properties:
        success:
          type: boolean
          const: false
        message:
          type: string
...
    AuthEnvelope:
      type: object
      required: [success, data]
      properties:
        success:
          type: boolean
          const: true
        data:
          $ref: '#/components/schemas/AuthResult'
...